refactor current code to server pacakge
