+++ spider/src/main/scala/dev/wishingtree/branch/spider/server/HttpVerb.scala
@@ -1,4 +1,6 @@
-package dev.wishingtree.branch.spider
+package dev.wishingtree.branch.spider.server
+
+import dev.wishingtree.branch.spider.server.HttpVerb
 
 import scala.util.Try
