+++ spider/src/test/scala/dev/wishingtree/branch/spider/HttpFunSuite.scala
@@ -2,8 +2,9 @@ package dev.wishingtree.branch.spider
 
 import com.sun.net.httpserver.HttpServer
 import dev.wishingtree.branch.lzy.LazyRuntime
+import dev.wishingtree.branch.spider.server.OpaqueSegments.*
+import dev.wishingtree.branch.spider.server.{ContextHandler, HttpVerb, RequestHandler}
 import munit.FunSuite
-import OpaqueSegments.*
 
 import java.net.InetSocketAddress
